# backend/docs/openapi/stashbox-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stashbox API",
    "version": "0.1.0",
    "description": "听匣（Stashbox）后端 4 服务合并 OpenAPI schema。\n\n由 `backend/scripts/export_openapi.py` 自动生成，请勿手改。\n\n服务分组见 tags：api-gateway / user-service / content-service / ai-service。"
  },
...
  "paths": {
...
    "/api/v1/admin/evaluations": {
      "get": {
        "summary": "Admin Evaluations List",
        "description": "distillation_evaluations 分页（评测员/用户/LLM 自评混排，按时间倒序）。",
        "operationId": "admin_evaluations_list_api_v1_admin_evaluations_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "auto_flag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Auto Flag"
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Score"
            }
          },
          {
            "name": "max_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Score"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "ai-service",
          "api-gateway"
        ]
      }
    },
...
    "/api/v1/articles/{article_id}/distill": {
      "post": {
        "summary": "Distill Article",
        "description": "开始蒸馏（v1 §3.2）：已在该文章上做过完整蒸馏则直接复用（按 article 幂等）；\n\n否则扣一次配额 → 建任务 → 入 Arq 队列。\n\n- CP1.7.4 幂等修复：复用判定从「已扣过配额」改为「已有完整蒸馏产物」\n  （不再把已在 content-service 抓取阶段扣过的配额误算入蒸馏阶段）\n- simulate_failure=True 用于验证「蒸馏失败 → 退还」\n- CP3.5-pre-3：任务不再在请求线程里跑（BackgroundTasks），而是塞进 Arq 队列由\n  独立 worker 进程消费；端点签名不变，只多返一个 job_id",
        "operationId": "distill_article_api_v1_articles__article_id__distill_post",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          },
          {
            "name": "simulate_failure",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Simulate Failure"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "ai-service",
          "api-gateway"
        ]
      }
    },
...
    "/api/v1/distill/start": {
      "post": {
        "summary": "Distill Start",
        "description": "按 article_id 触发蒸馏。\n\nCP-DISTILL-START-HARDEN（本轮修复的三个问题，都是实测出来的）：\n\n1) **越权**。原实现只信请求体里的 article_id/url，从不查 Article，\n   任何人拿到别人的 article_id 就能对其重跑蒸馏。实测：以 user 9277 的\n   token 对 user 1 名下的 art_wxSOP_verify_0001 调本端点返回 200。\n   现在查库校验归属，非本人 → 403（与 /articles/{id}/distill 同口径）。\n\n2) **不扣配额**。本端点从头到尾没碰过 quota_service，而\n   /articles/{id}/distill 扣。同一件事两条路径、两种计费口径。\n\n3) **破坏性复用**。文章已有蒸馏行时复用原行，并把 script_text / audio_url /\n   duration_sec / quality_score / tags 一律置空。实测把一篇 status=done、\n   有音频的真机在用文章打成了 status=queued，产物元数据全部丢失，且无法找回。\n   改为**保留旧产物**：重跑期间用户仍能听旧音频，worker 跑成功才覆盖。\n\nurl / title 一律取库里的值，不用请求体传的——传什么蒸馏什么，等于允许\n调用方把 A 文章的产物写到 B 上。\n\nCP-TTS-VOICE：**本端点同时是「用当前音色重新生成」的入口**，不另开新端点。\n理由是本项目已经吃过亏 —— 见上面第 2) 条：「同一件事两条路径、两种计费口径」。\n再加一个 `/re-distill` 只会重演：它要么不扣费（和这里不一致），要么重复扣费\n（用户点两次就少两次配额）。音色是合成期参数，worker 里\n`resolve_voice_for_user()` 会重新解析用户当前选的音色，所以重跑自然生效。",
        "operationId": "distill_start_api_v1_distill_start_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistillStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistillStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "ai-service",
          "api-gateway"
        ]
      }
    },
    "/api/v1/distill/{task_id}": {
      "get": {
        "summary": "Distill Status",
        "operationId": "distill_status_api_v1_distill__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "ai-service",
          "api-gateway"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "DistillStartRequest": {
        "properties": {
          "article_id": {
            "type": "string",
            "title": "Article Id"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "article_id"
        ],
        "title": "DistillStartRequest"
      },
      "DistillStartResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "article_id": {
            "type": "string",
            "title": "Article Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "default": ""
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Id"
          },
          "voice_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Name"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "article_id",
          "status"
        ],
        "title": "DistillStartResponse"
      },
...
    },
    "securitySchemes": {}
  }
}